pyproject.toml
[project]
name = "haystack-data-convertor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11, <3.12"
dependencies = [
    "colorlog>=6.9.0",
    "lxml>=6.0.0",
    "pypdf>=6.0.0",
    "pydantic>=2.0",
    "pytest>=8.4.1",
    "qdrant-client>=1.15.1",
    "ruff>=0.12.8",
    "sentence-transformers>=5.1.0",
    "unstructured>=0.18.11",
    "unstructured-fileconverter-haystack>=0.5.0",
    "haystack-ai>=2.16.1",
    "qdrant-haystack>=9.2.0",
    "python-dotenv>=1.1.1",
    "gradio>=5.42.0",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.30",
    "markdown>=3.8.2",
    "pymupdf>=1.26.3",
    "pdfplumber>=0.11.7",
    "python-docx>=1.2.0",
    "pytesseract>=0.3.13",
    "pdfminer-six>=20250506",
]
